+++ fabfile.py
@@ -1,6 +1,6 @@
 from fabric.api import env, local, prefix, put, sudo, task, \
      require, puts, cd, run, abort
-from fabric.colors import green, red, cyan
+from fabric.colors import green, red, cyan, yellow
 from fabric.contrib import files
 from fabric.context_managers import cd, hide, settings
 import eulindexer
@@ -24,6 +24,7 @@ that you are interested in.
 
 # default settings
 env.rev_tag = ''
+env.project = 'eulindexer'
 env.remote_path = '/home/httpd/sites/eulindexer'
 env.remote_acct = 'eulindexer'
 env.url_prefix = ''
@@ -34,8 +35,8 @@ def configure(path=None, user=None, url_prefix=None, remote_proxy=None):
     env.version = eulindexer.__version__
     config_from_git()
     # construct a unique build directory name based on software version and svn revision
-    env.build_dir = 'eulindexer-%(version)s%(rev_tag)s' % env
-    env.tarball = 'eulindexer-%(version)s%(rev_tag)s.tar.bz2' % env
+    env.build_dir = '%(project)s-%(version)s%(rev_tag)s' % env
+    env.tarball = '%(project)s-%(version)s%(rev_tag)s.tar.bz2' % env
 
     if path:
         env.remote_path = path.rstrip('/')
@@ -111,31 +112,55 @@ def setup_virtualenv():
              user=env.remote_acct)
         # activate the environment and install required packages
         with prefix('source env/bin/activate'):
+            # required packages
             pip_cmd = 'pip install -r pip-install-req.txt'
             if env.remote_proxy:
                 pip_cmd += ' --proxy=%(remote_proxy)s' % env
             sudo(pip_cmd, user=env.remote_acct)
+            # install optional modules (progressbar)
+            pip_cmd = 'pip install -r pip-install-opt.txt'
+            if env.remote_proxy:
+                pip_cmd += ' --proxy=%(remote_proxy)s' % env
+            sudo(pip_cmd, user=env.remote_acct)
+                        
 
 def configure_site():
     'Copy configuration files into the remote source tree.'
     with cd(env.remote_path):
         if not files.exists('localsettings.py'):  
             abort('Configuration file is not in expected location: %(remote_path)s/localsettings.py' % env)
-        sudo('cp localsettings.py %(build_dir)s/eulindexer/localsettings.py' % env,
+        sudo('cp localsettings.py %(build_dir)s/%(project)s/localsettings.py' % env,
              user=env.remote_acct)
 
     # collect static files to be served out by apache
     with cd('%(remote_path)s/%(build_dir)s' % env):
         with prefix('source env/bin/activate'):
-            sudo('python eulindexer/manage.py collectstatic --noinput',
+            sudo('python %(project)s/manage.py collectstatic --noinput' % env,
                  user=env.remote_acct)
 
 def update_links():
     'Update current/previous symlinks on the remote server.'
     with cd(env.remote_path):
-        if files.exists('current' % env):
+        if files.exists('current'):
             sudo('rm -f previous; mv current previous', user=env.remote_acct)
         sudo('ln -sf %(build_dir)s current' % env, user=env.remote_acct)
+
+@task
+def compare_localsettings(path=None, user=None):
+    'Compare current/previous (if any) localsettings on the remote server.'
+    configure(path, user)
+    with cd(env.remote_path):
+        # sanity-check current localsettings against previous
+        if files.exists('previous'):
+            with settings(hide('warnings', 'running', 'stdout', 'stderr'),
+                          warn_only=True):  # suppress output, don't abort on diff error exit code
+                output = sudo('diff current/%(project)s/localsettings.py previous/%(project)s/localsettings.py' % env,
+                              user=env.remote_acct)
+                if output:
+                    puts(yellow('WARNING: found differences between current and previous localsettings.py'))
+                    puts(output)
+                else:
+                    puts(green('No differences between current and previous localsettings.py'))
 
 
 def build_source_package(path=None, user=None, url_prefix='', remote_proxy=''):
@@ -170,6 +195,7 @@ def deploy(path=None, user=None, url_prefix='', remote_proxy=''):
     setup_virtualenv()
     configure_site()
     update_links()
+    compare_localsettings()
 
     puts(green('Successfully deployed %(build_dir)s to %(host)s' % env))
 
@@ -177,13 +203,14 @@ def deploy(path=None, user=None, url_prefix='', remote_proxy=''):
 def revert(path=None, user=None):
     """Update remote symlinks to retore the previous version as current"""
     configure(path, user)
-    # if there is a previous link, shift current to previous
-    if files.exists('previous'):
-        # remove the current link (but not actually removing code)
-        sudo('rm current', user=env.remote_acct)
-        # make previous link current
-        sudo('mv previous current', user=env.remote_acct)
-        sudo('readlink current', user=env.remote_acct)
+    with cd(env.remote_path):
+        # if there is a previous link, shift current to previous
+        if files.exists('previous'):
+            # remove the current link (but not actually removing code)
+            sudo('rm current', user=env.remote_acct)
+            # make previous link current
+            sudo('mv previous current', user=env.remote_acct)
+            sudo('readlink current', user=env.remote_acct)
 
 @task
 def clean():
@@ -211,7 +238,7 @@ def rm_old_builds(path=None, user=None, noinput=False):
         dir_items = [n.strip() for n in dir_listing.split('\n')] 
         # regex based on how we generate the build directory:
         #   project name, numeric version, optional pre/dev suffix, optional revision #
-        build_dir_regex = r'^eulindexer-[0-9.]+(-[A-Za-z0-9_-]+)?(-r[0-9a-f]+)?$' % env
+        build_dir_regex = r'^%(project)s-[0-9.]+(-[A-Za-z0-9_-]+)?(-r[0-9a-f]+)?$' % env
         build_dirs = [item for item in dir_items if re.match(build_dir_regex, item)]
         # by default, preserve the 3 most recent build dirs from deletion
         rm_dirs = build_dirs[3:]